backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
Pillow==10.0.0
requests==2.31.0
motor==3.3.2
pymongo==4.6.0